+++ Packages/com.unity.inputsystem/InputSystem/Actions/InputActionState.cs
@@ -1129,6 +1129,14 @@ namespace UnityEngine.Experimental.Input
                 memory.controlMagnitudes[trigger.controlIndex] = trigger.magnitude;
             }
 
+            // Never ignore state changes for actions that aren't currently driven by
+            // anything.
+            if (actionState->controlIndex == kInvalidIndex)
+            {
+                Profiler.EndSample();
+                return false;
+            }
+
             // If the control is actuated *more* than the current level of actuation we recorded for the
             // action, we process the state change normally. If this isn't the control that is already
             // driving the action, it will become the one now.
@@ -1657,6 +1665,14 @@ namespace UnityEngine.Experimental.Input
                         RemoveContinuousAction(actionIndex);
                     break;
                 }
+            }
+
+            // If we're now waiting, reset control state. This is important for the disambiguation code
+            // to not consider whatever control actuation happened on the action last.
+            if (actionState->phase == InputActionPhase.Waiting)
+            {
+                actionState->controlIndex = kInvalidIndex;
+                actionState->flags &= ~TriggerState.Flags.HaveMagnitude;
             }
         }